flet_frontend/requirements.txt
flet>=0.21.0
httpx>=0.27.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.7
pillow>=10.2.0
face-recognition>=1.3.0
aiosqlite>=0.20.0  # For async SQLite operations
pytest>=8.0.0
pytest-asyncio>=0.23.0  # For async tests